Move SecretTextArea and TextInputContainer to components/ui
