+++ airbyte-webapp/src/components/ui/SecretTextArea/SecretTextArea.module.scss
