+++ airbyte-webapp/src/components/ui/SecretTextArea/SecretTextArea.test.tsx
