+++ airbyte-webapp/src/components/ui/SecretTextArea/SecretTextArea.tsx
