+++ airbyte-webapp/src/components/ui/SecretTextArea/index.ts
